# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Basic CRUD Laravel API Documentation",
        "description": "API Documentation Test Flix",
        "contact": {
            "email": "jauharimalikupil@gmail.com"
        },
        "license": {
            "name": "GPL2",
            "url": "https://jauharimalik.github.io"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "jauharimalikupil@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register User",
                "description": "Register New User",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jauhari Malik"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "jauharimalikupil@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register New User Data"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticated User Profile",
                "description": "Authenticated User Profile",
                "operationId": "b9abb1a7a74670a19c215c2c133f14d8",
                "responses": {
                    "200": {
                        "description": "Authenticated User Profile"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Logout"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh",
                "description": "Refresh",
                "operationId": "b2eae7a8032954d25b46c7ec6f489c68",
                "responses": {
                    "200": {
                        "description": "Refresh"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/movies": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get list of movies",
                "description": "Returns list of movies",
                "operationId": "index movies",
                "responses": {
                    "200": {
                        "description": "Get Movies List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Movies"
                ],
                "summary": "Create a new movie",
                "description": "Creates a new movie record",
                "operationId": "createMovie",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Pengabdi Setan 2 Comunion"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "dalah sebuah film horor Indonesia tahun 2022 yang disutradarai dan ditulis oleh Joko Anwar sebagai sekuel dari film tahun 2017, Pengabdi Setan."
                                    },
                                    "rating": {
                                        "type": "number",
                                        "example": 7
                                    },
                                    "kode": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movie created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/movies/{movie}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get a movie by ID",
                "description": "Returns a movie by its ID",
                "operationId": "getMovieById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Movies Details"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Movies"
                ],
                "summary": "Delete Movie",
                "description": "Delete Movie from Database",
                "operationId": "deleteMovie",
                "parameters": [
                    {
                        "name": "movie",
                        "in": "path",
                        "description": "ID of the movie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Movie deleted successfully"
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Movies"
                ],
                "summary": "Update Movie",
                "description": "Updates a movie by its ID",
                "operationId": "updateMovie",
                "parameters": [
                    {
                        "name": "movie",
                        "in": "path",
                        "description": "ID of the movie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Pengabdi Setan 2 Comunion"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "dalah sebuah film horor Indonesia tahun 2022 yang disutradarai dan ditulis oleh Joko Anwar sebagai sekuel dari film tahun 2017, Pengabdi Setan."
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 7
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "kode": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "created_at": {
                                        "type": "string",
                                        "example": "2022-08-01 10:56:31"
                                    },
                                    "updated_at": {
                                        "type": "string",
                                        "example": "2022-08-13 09:30:23"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Movie updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                }
            }
        }
    }
}